move isUrl validator to BusinessForm given it is specific to that form
