+++ static/js/src/mixins/form/starter.js
@@ -85,14 +85,6 @@ export default {
       return typeof isValid === 'undefined';
     },
 
-    isURL(value) {
-      const isValid = validate(
-        { website: value.trim() },
-        { website: { url: true } },
-      );
-      return typeof isValid === 'undefined';
-    },
-
     isNumeric(value) {
       const isValid = validate(
         { value: value.trim() },
